papersqueeze/requirements.txt
# PaperSqueeze Dependencies

# HTTP client
httpx>=0.27.0

# Configuration
pyyaml>=6.0
pydantic>=2.0

# Future: LLM providers (commented out for MVP)
# anthropic>=0.39.0
# openai>=1.0.0
